backend/requirements.txt
# OmniPublish V2.0 — Python 依赖
# 核心框架
fastapi==0.115.0
uvicorn[standard]==0.30.0

# 数据库
aiosqlite==0.20.0

# 认证
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3                     # passlib 不兼容 bcrypt 5.x

# 图片处理
Pillow>=10.0.0
numpy>=1.24.0

# 视频处理（YOLOv8 人脸检测）
# ultralytics>=8.0.0  # 可选，首次运行会自动下载模型

# CMS 发布
requests>=2.28.0
pycryptodome>=3.18.0

# 开发工具
python-multipart>=0.0.9  # 文件上传支持

# 测试
pytest>=8.0.0
httpx>=0.27.0             # FastAPI TestClient 依赖
